reallocated logic
created seperate class for moving logic and moved allocated code from move to this class
